+++ package.json
@@ -1,15 +1,11 @@
 {
   "private": true,
   "sideEffects": false,
+  "type": "module",
   "scripts": {
-    "build": "run-s build:*",
-    "build:remix": "remix build",
-    "build:server": "esbuild --platform=node --format=cjs ./server.ts --outdir=build --bundle",
-    "dev": "run-p dev:*",
-    "dev:build": "cross-env NODE_ENV=development npm run build:server -- --watch",
-    "dev:remix": "cross-env NODE_ENV=development remix watch",
-    "dev:server": "cross-env NODE_ENV=development node --inspect --require dotenv/config ./build/server.js",
-    "start": "cross-env NODE_ENV=production node ./build/server.js",
+    "build": "remix build",
+    "dev": "remix dev --no-restart -c \"node server.js\"",
+    "start": "cross-env NODE_ENV=production node ./server.js",
     "typecheck": "tsc"
   },
   "dependencies": {
@@ -36,10 +32,12 @@
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.8",
     "autoprefixer": "^10.4.14",
+    "chokidar": "^3.5.3",
     "dotenv": "^16.0.3",
     "esbuild": "^0.17.16",
     "eslint": "^8.27.0",
     "npm-run-all": "^4.1.5",
+    "postcss-load-config": "^4.0.1",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.2.7",
     "tailwindcss": "^3.3.1",
