+++ tsconfig.json
@@ -16,7 +16,6 @@
     "paths": {
       "~/*": ["./app/*"]
     },
-    "types": ["node"],
 
     // Remix takes care of building everything in `remix build`.
     "noEmit": true
